requirements.txt
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.3.0
Aspose.Slides==24.5.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
coloredlogs==15.0.1
cryptography==42.0.7
datrie==0.8.2
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.3.1
h11==0.14.0
hanziconv==0.3.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
nltk==3.8.1
numpy==1.26.4
onnxruntime==1.16.3
opencv-python==4.9.0.80
orjson==3.10.3
packaging==24.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
protobuf==5.26.1
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.10
requests==2.27.1
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.13.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
StrEnum==0.4.15
sympy==1.12
tiktoken==0.6.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
